examples/pytorch_lightning/requirements.txt
bitorch
fvbitcore
pytorch_lightning
sklearn
wandb
